Infinite Thank You messages
Got some of the functionality working but not all of it.
Adding the carriage return and newline characters was successful.

However, the overall input was not as blocking as I thought it would
be.  Maybe I should switch back to bus-pirate to fine-tune the input
methods.

Also, remember to use volatile in 3 scenarios
 - when accessing a mem-mapped peripheral
 - when sharing variable with interrupt handlers
 - when sharing global variables between different threads