Readln/writeln done. Start menu, bootldr
Completed base functionality for readln and writeln.
Typed characters will echo back to you.  Newlines and
carriage returns are used to separate text.  We no longer
have an infinite number of Thank Yous.  That was fixed by not
accepting the null character and polling for characters inside
the while loop.

Began work on a menu interface.  Plan to include the ability to
load code over UART.  Also created some basic utilities.  None
of this functionality is even close to being ready for testing.

Need to investigate how to store an array of strings as a global
constant.