Add bootldr.h. Branch Command Pkt in prograspi.c
Create a header file for all the bootldr structs and enums.  This
way they can be shared between client and server.

Re-worked the main method in prograspi.c so that the branch command
uses the new bootldr struct.  Still need to rework reading the hex
file and sending the data over the serial line.

Have not tried to compile.